requirements.txt
Jinja2==2.9.5
Mako==1.0.2
MarkupSafe==1.1.1
WebOb==1.8.6
WebTest==2.0.18
beautifulsoup4==4.3.2
flake8
ipdb==0.8.1
ipython
mccabe
pecan
pep8
py
pyflakes
pytest
requests==2.8.0
singledispatch==3.4.0.2
six
waitress==0.8.7
pytz
tzlocal
CherryPy==3.8.0
python-dotenv
oauthlib
Beaker
requests-oauthlib==1.3.0
